start of RTIOWK

+++ crates/rtow/src/primitives.rs
@@ -7,7 +7,7 @@ mod vec3;
 pub use ray::Ray;
 pub use vec3::Vec3;
 
-pub type Decimal = f32;
+pub type Decimal = f64;
 pub mod decimal_consts {
     pub use std::f32::consts::*;
 }
